ai.devictoria.shop/services/authservice/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic-settings==2.6.1
PyJWT==2.10.1
httpx==0.28.1
redis==7.1.0

